==> notebooks/01_admin_boundary_processing.ipynb <==
import os

os.chdir("..")   # move out of notebooks/
print(os.getcwd())


os.listdir("gis/raw")


from pyrosm import OSM
import geopandas as gpd
import matplotlib.pyplot as plt

plt.rcParams["figure.figsize"] = (8,8)


# Delhi NCR approx bounding box
# (west, south, east, north)
delhi_ncr_bbox = [76.5, 28.2, 77.7, 29.0]


osm = OSM(
    "gis/raw/northern-zone-260116.osm.pbf",
    bounding_box=delhi_ncr_bbox
)






boundaries = osm.get_boundaries()
boundaries.head()


boundaries["admin_level"].value_counts()


delhi_ncr = boundaries[
    (boundaries["admin_level"] == 6) &
    (boundaries["name"].str.contains(
        "Delhi|New Delhi|Central Delhi|North Delhi|South Delhi|East Delhi|West Delhi|"
        "Ghaziabad|Noida|Gurgaon|Gurugram|Faridabad",
        case=False,
        na=False
    ))
]


#ax = delhi_ncr.plot(edgecolor="black", figsize=(8,8))
#ax.set_aspect("auto")


delhi_ncr = delhi_ncr.to_crs(epsg=4326)

delhi_ncr.to_file(
    "gis/processed/delhi_ncr_districts.geojson",
    driver="GeoJSON"
)


import os
os.path.exists("gis/processed/delhi_ncr_districts.geojson")
