==> notebooks/02_road_network_processing.ipynb <==
import os

os.chdir(r"C:\Users\hp\Documents\ML_Projects\Delhi_Infrastructure_Risk_AI")
print(os.getcwd())



os.listdir()


from pyrosm import OSM
import geopandas as gpd
from shapely.ops import unary_union
import os



boundary = gpd.read_file("gis/processed/delhi_ncr_districts.geojson")
boundary = boundary.to_crs(epsg=4326)



minx, miny, maxx, maxy = boundary.total_bounds

bbox = [minx, miny, maxx, maxy]
bbox


bbox = [76.5, 28.2, 77.7, 29.0]   # Delhi NCR bbox


print(bbox)
print(type(bbox))
all(isinstance(x, (int, float)) for x in bbox)


from pyrosm import OSM

osm = OSM(
    "gis/raw/northern-zone-260116.osm.pbf",
    bounding_box=bbox
)



roads = osm.get_network(network_type="driving")
roads.head()


roads = roads[
    ["name", "highway", "bridge", "tunnel", "length", "geometry"]
]


roads["is_flyover"] = roads["bridge"].notna()
roads["is_flyover"].value_counts()


roads.columns


roads = gpd.overlay(roads, boundary, how="intersection")


roads = gpd.sjoin(
    roads,
    boundary[["geometry"]],
    how="inner",
    predicate="intersects"
)

roads = roads.drop(columns=["index_right"])


roads.to_file(
    "gis/processed/delhi_ncr_roads_flyovers.geojson",
    driver="GeoJSON"
)


import os
os.path.exists("gis/processed/delhi_ncr_roads_flyovers.geojson")
