==> notebooks/05_flood_impact_model.ipynb <==
import os
os.chdir(r"C:\Users\hp\Documents\ML_Projects\Delhi_Infrastructure_Risk_AI")
print(os.getcwd())


import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error, r2_score


df = pd.read_csv("data/processed/engineered_features.csv")
df.head()


df["flood_severity"] = (
    0.45 * df["FEI"] +
    0.25 * df["TSI"] +
    0.20 * df["SAI"] +
    0.10 * df["CDI"]
)

df["flood_severity"] = (
    df["flood_severity"] / df["flood_severity"].max()
)



X = df[["FEI", "TSI", "SAI", "CDI", "is_flyover"]]
y = df["flood_severity"]


X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.25,
    random_state=42
)


flood_model = RandomForestRegressor(
    n_estimators=300,
    max_depth=8,
    random_state=42
)

flood_model.fit(X_train, y_train)


y_pred = flood_model.predict(X_test)

print("MAE:", mean_absolute_error(y_test, y_pred))
print("R2 Score:", r2_score(y_test, y_pred))


importance = pd.DataFrame({
    "Feature": X.columns,
    "Importance": flood_model.feature_importances_
}).sort_values(by="Importance", ascending=False)

importance


df["predicted_flood_impact"] = flood_model.predict(X)
df[["asset_id", "predicted_flood_impact"]].head()


df.to_csv(
    "data/processed/flood_impact_predictions.csv",
    index=False
)


import joblib

joblib.dump(
    flood_model,
    "models/flood_impact_model.pkl"
)
