==> notebooks/06_priority_index_calculation.ipynb <==
import os
os.chdir(r"C:\Users\hp\Documents\ML_Projects\Delhi_Infrastructure_Risk_AI")
print(os.getcwd())


import pandas as pd
import numpy as np


risk_df = pd.read_csv("data/processed/risk_predictions.csv")
flood_df = pd.read_csv("data/processed/flood_impact_predictions.csv")


risk_df.shape, flood_df.shape


df = risk_df.merge(
    flood_df[["asset_id", "predicted_flood_impact"]],
    on="asset_id",
    how="left"
)

df.head()


def normalize(col):
    return (col - col.min()) / (col.max() - col.min())

df["risk_score_norm"] = normalize(df["risk_score"])
df["flood_impact_norm"] = normalize(df["predicted_flood_impact"])


df["D_CIPI"] = (
    0.40 * df["risk_score_norm"] +
    0.30 * df["flood_impact_norm"] +
    0.20 * df["MNS"] +
    0.10 * df["TSI"]
)


df["priority_level"] = pd.cut(
    df["D_CIPI"],
    bins=[0, 0.4, 0.7, 1.0],
    labels=["Low", "Medium", "High"]
)

df["priority_level"].value_counts()


df = df.sort_values(by="D_CIPI", ascending=False)

df[["asset_id", "D_CIPI", "priority_level"]].head(10)


df.to_csv(
    "data/processed/infrastructure_priority_index.csv",
    index=False
)
